refactor(sql-editor): rename sql-editor dir to SqlEditor
