+++ frontend/src/views/SqlEditor/EditorHeader.vue
