+++ frontend/src/views/SqlEditor/QuerySelector.vue
